light refactoring
